implement DateTime::secondstime() to sidestep the Y2036 NTP wrap comparison
gps_serial_poll() rejects a GPS fix as "bad" by comparing gpstime
against compileTime via ntptime(), which is deliberately wire-format-
constrained and wraps at 2036-02-07. Once the real GPS date crosses
that wrap, the comparison flips permanently and the firmware rejects
every real fix from then on, with no recovery short of a rebuild.

secondstime() (declared but never implemented) has no such
constraint -- it's a plain seconds-since-2000 count, monotonic out to
~2136 -- so this wires it in for that one comparison only, leaving
ntptime() and everything downstream (ClockPID/NTPClock/NTPServer)
untouched. Also widens the declared return type from signed long to
uint32_t, since a signed 32-bit counter would only stay monotonic for
~68 years, not the ~136 the design calls for, and secondstime() had
no existing callers to break.

Also moves the now-completed "NTP timestamp era rollover (Y2036)"
TODO.md item into DONE.md, matching every other closed-out item this
session.

+++ DateTime.h
@@ -22,8 +22,13 @@ class DateTime {
   uint16_t minute() const { return minute_; }
   uint16_t second() const { return second_; }
 
-  // 32-bit times as seconds since 1/1/2000
-  long secondstime() const;
+  // 32-bit times as seconds since 1/1/2000, unsigned so it's monotonic with
+  // real time all the way out to ~2136 (2^32 seconds from the 2000 epoch) --
+  // no NTP-wire-format constraint forcing it to alias onto 1900-2000 the way
+  // ntptime() deliberately does, so plain `<` works for comparing two
+  // firmware-internal timestamps without a wraparound-aware helper. See
+  // TODO.md/DONE.md, "NTP timestamp era rollover (Y2036)".
+  uint32_t secondstime() const;
   // 32-bit times as seconds since 1/1/1900, TAI-like: real, leap-second-
   // adjusted seconds, monotonic across a leap second insertion/deletion --
   // NOT the raw NTP wire format. Convert via LeapSeconds.h's
